Rename ramadan-activities to ramadan
