+++ i18n.json
@@ -58,7 +58,7 @@
     "/login": ["login"],
     "/about-the-quran": ["about-quran"],
     "/notes-and-reflections": ["notes"],
-    "/ramadan-activities": ["ramadan-activities"],
+    "/ramadan": ["ramadan-activities"],
     "/notification-settings": ["notification-settings"],
     "/calendar": ["quranic-calendar"],
     "/media": ["media"],
