+++ src/pages/what-is-ramadan/index.tsx
@@ -10,7 +10,7 @@ import PageContainer from '@/components/PageContainer';
 import InlineLink from '@/components/RamadanActivity/InlineLink';
 import { getWhatIsRamadanOgImageUrl } from '@/lib/og';
 import styles from '@/pages/contentPage.module.scss';
-import pageStyles from '@/pages/ramadan-activities/RamadanActivities.module.scss';
+import pageStyles from '@/pages/ramadan/RamadanActivities.module.scss';
 import { getLanguageAlternates } from '@/utils/locale';
 import { getCanonicalUrl, getWhatIsRamadanNavigationUrl } from '@/utils/navigation';
