+++ src/utils/navigation.ts
@@ -346,7 +346,7 @@ export const getCollectionNavigationUrl = (collectionId: string) => {
 export const getReadingGoalNavigationUrl = () => '/reading-goal';
 export const getMyCoursesNavigationUrl = () => '/my-learning-plans';
 export const getCoursesNavigationUrl = () => '/learning-plans';
-export const getRamadanActivitiesNavigationUrl = () => '/ramadan-activities';
+export const getRamadanNavigationUrl = () => '/ramadan';
 export const getWhatIsRamadanNavigationUrl = () => '/what-is-ramadan';
 export const getLoginNavigationUrl = (redirectTo?: string) =>
   `/login${redirectTo ? `?${QueryParam.REDIRECT_TO}=${redirectTo}` : ''}`;
